refactor(datasets): extract docs/journal/notifications/thumbnail/admin routes to routes/misc.ts (6d)
Last route group out of router.js (now 33 lines: global resourceType .use + register calls).
Typed locks collection (<{_id:string}>) for the string-keyed lock docs. Preserved
bit-for-bit: user-notification passes a stray 'true' in notifications.send's sessionState
position (parking lot).
